--- Mamífero.java
@@ -1,91 +1,0 @@
-
-
-package exercícioaula02;
-
-/**
- *
- * @author Vitória Stefany
- */
-public class Mamífero {
-    private String habitat;
-    private String alimentacao;
-    private String qtdPatas;
-    private boolean nadadeiras;
-    private String cor;
-    private int espectativaVida;
-    private boolean dente;
-    private boolean tamanho;
-
-    public String getHabitat() {
-        return habitat;
-    }
-
-    public void setHabitat(String habitat) {
-        this.habitat = habitat;
-    }
-
-    public String getAlimentacao() {
-        return alimentacao;
-    }
-
-    public void setAlimentacao(String alimentacao) {
-        this.alimentacao = alimentacao;
-    }
-
-    public boolean isQtdPatas() {
-        return qtdPatas;
-    }
-
-    public void setQtdPatas(boolean qtdPatas) {
-        this.qtdPatas = qtdPatas;
-    }
-
-    public boolean isNadadeiras() {
-        return nadadeiras;
-    }
-
-    public void setNadadeiras(boolean nadadeiras) {
-        this.nadadeiras = nadadeiras;
-    }
-
-    public String getCor() {
-        return cor;
-    }
-
-    public void setCor(String cor) {
-        this.cor = cor;
-    }
-
-    public String getQtdPatas() {
-        return qtdPatas;
-    }
-
-    public void setQtdPatas(String qtdPatas) {
-        this.qtdPatas = qtdPatas;
-    }
-
-    public int getEspectativaVida() {
-        return espectativaVida;
-    }
-
-    public void setEspectativaVida(int espectativaVida) {
-        this.espectativaVida = espectativaVida;
-    }
-
-    public boolean isDente() {
-        return dente;
-    }
-
-    public void setDente(boolean dente) {
-        this.dente = dente;
-    }
-
-    public boolean isTamanho() {
-        return tamanho;
-    }
-
-    public void setTamanho(boolean tamanho) {
-        this.tamanho = tamanho;
-    }
-    
-}
